Change in Chatbot
Due to limitations and complications with OpenAI's API, we decided to stick to a Machine Learning model and implement a suggested questions for users so they can get an idea of what they can ask/what the bot can do. If it is a question that the bot has never seen, it will add it to a list of unanswered questions. Also has some autocorrecting